Move link; Align tops of columns
